Moved RTMockClient into routingtable package to make it reusable.
